Changing to auto_load for properties and classes
Utilises the auto_load.py for auto-register the classes and properties.
This leads to a cleaner __init__.py without having to manually specify
the register_class() and unregister_class() for each class, the same
with deleting the properties.

This will enable moving the specific properties to more relevant
submodules, enabling a cleaner organisation of the addon.