fixed density importing
it seems that there is now a blender 4.1 issue where even if the underlying `.vdb` changes, if you reimport the file it still uses the preview for the 'old' version.

Because of this, the created .vdb file path now contains whether or not it was centred and inverted, which makes different files and fixes the issues.