Adding tools for replicating MD trajectories across periodic boundaries (#1035)
This PR attempts to address Issue #1026. It adds 4 new nodes:

Periodic Box - this node takes as input the a, b, c lengths and alpha, beta, gamma angles to define a simulation box and outputs the corresponding vectors. This node updates with the trajectory to allow for boxes that change during the simulation. The parameters can also be entered manually for use with non-MD inputs e.g. pdb structures
Lattice Grid - this node takes the box vectors as input and provides a grid of lattice points. By inputting these points into a instance on points node the system can be replicated at each of these points.
Periodic Image - this node provides the location of a single lattice point to allow the duplication of the system at a specific point/set of points
Replicate Trajectory - this node bundles up a Periodic Box, Lattice Grid and Instance on Points node to allow for replication of a trajectory across a grid with one node.

---------

Co-authored-by: Brady Johnston <brady.johnston@me.com>